+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Installation - Fix installation errors due to express-cors package using Yarn.
+- Exports - Fix bad initial implementation for exports authentication.
 
 ## RELEASE 1.2.0 - 2017-08-21
 ### Added